searching somewhat works now (but results are underwhelming)

+++ src/main/java/score/ScoreUtility.java
@@ -16,14 +16,15 @@ public class ScoreUtility {
 
     public double tf(InvertedIndex index, DocumentInfo documentInfo, String word) {
         IndexValue indexValue = index.findByWord(word);
-        double frequency = (double) indexValue.getFrequencyInDocument(documentInfo.getId());
-        double maxFrequency = (double) documentInfo.getMaxFrequencyOfWord();
+        double frequency = indexValue.getFrequencyInDocument(documentInfo.getId());
+        double maxFrequency = documentInfo.getMaxFrequencyOfWord();
         return frequency / maxFrequency;
     }
 
     public double idf(InvertedIndex index, String word) {
         double numDocumentsContainingWord = index.findByWord(word).getNumDocuments();
-        double numDocuments = (double) index.getDocumentRepository().getNumDocuments();
-        return Math.log(numDocuments / numDocumentsContainingWord); // according to "An Introduction to IR" page 155
+        double numDocuments = index.getDocumentRepository().getNumDocuments();
+        // according to "An Introduction to IR" page 155
+        return Math.log(numDocuments / numDocumentsContainingWord);
     }
 }
